classifier

+++ efficiencies/classifier/data.py
@@ -10,7 +10,7 @@ class isReco_Dataset(torch.utils.data.Dataset):
             h5py_file["GenElectrons"][start : (start + stop), 0:38], dtype=torch.float32
         )
         self.y = torch.tensor(
-            h5py_file["GenElectrons"][start : (start + stop), -1], dtype=torch.long
+            h5py_file["GenElectrons"][start : (start + stop), -1], dtype=torch.float32
         ).view(-1, 1)
 
     def __len__(self):
